% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HH2_soil_moisture_import.R
\name{HH2Import}
\alias{HH2Import}
\title{Import function for Delta-T Devices HH2 Soil moisture files}
\usage{
HH2Import(filename, sensor.type)
}
\arguments{
\item{filename}{The name of the file to be imported. Character string.}

\item{sensor.type}{Either \code{"PR2"} for soil mositure profile probe or \code{"ML3"} for TDR. Other values are not allowed.}
}
\value{
Returns a data frame with the elements Time, Sample, Plot, Device, Depth, Percent_Vol, Vol_Error, mV, mV_Error.
}
\description{
Imports a file created by the DeltaT handheld soil mositure reader (HH2). Works for profile probes (\code{PR2}) or TDRs (\code{ML3}) files. Sensor type must be specified at import. Re-organises the data of the HH2 file to be compatible with a data frame.
}

